- share model between test and program, add typeallias for hash, add constant for EMPTY_HASH
