+++ arch/contracts/exchange/Cargo.toml
@@ -15,6 +15,7 @@ log = "0.4.22"
 env_logger = "0.11.5"
 lazy_static = "1.5.0"
 ctor = { version = "0.2.8", features = [] }
+model = { path = "model" }
 
 [dev-dependencies]
 serial_test = "3.1.1"
