+++ arch/contracts/exchange/model/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "model"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+borsh = { version = "1.5.1", features = ["derive"] }
+arch_program = { path = "../../../program" }
