+++ arch/contracts/exchange/model/src/error.rs
@@ -1,0 +1,15 @@
+
+pub const ERROR_INVALID_ADDRESS_INDEX: u32 = 601;
+pub const ERROR_INVALID_ACCOUNT_INDEX: u32 = 602;
+pub const ERROR_INSUFFICIENT_BALANCE: u32 = 603;
+pub const ERROR_ADDRESS_MISMATCH: u32 = 604;
+pub const ERROR_SETTLEMENT_IN_PROGRESS: u32 = 605;
+pub const ERROR_NO_SETTLEMENT_IN_PROGRESS: u32 = 606;
+pub const ERROR_SETTLEMENT_BATCH_MISMATCH: u32 = 607;
+pub const ERROR_NETTING: u32 = 608;
+pub const ERROR_ALREADY_INITIALIZED: u32 = 609;
+pub const ERROR_PROGRAM_STATE_MISMATCH: u32 = 610;
+pub const ERROR_NO_OUTPUTS_ALLOWED: u32 = 611;
+pub const ERROR_INVALID_ADDRESS: u32 = 612;
+pub const ERROR_INVALID_SIGNER: u32 = 613;
+pub const ERROR_VALUE_TOO_LONG: u32 = 613;
