+++ arch/contracts/exchange/model/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod exchange;
+pub mod error;
