+++ arch/contracts/exchange/program/Cargo.toml
@@ -12,6 +12,7 @@ borsh = { version = "1.5.1", features = ["derive"] }
 sha256 = "1.5.0"
 bitcoin = { version = "0.31.0", features = ["serde"] }
 hex = { version = "0.4.3", default-features = false }
+model = { path = "../model" }
 
 [lib]
 crate-type = ["cdylib", "lib"]
